AVRO-4295: [csharp] Cast bounded length to int; clamp structural cap
Addresses review feedback:
 - read(long p) now allocates new byte[(int)p]; p is already bounded to
   <= MaxDotNetArrayLength above, so the cast required for array allocation
   cannot overflow.
 - MaxCollectionStructural is clamped to int.MaxValue. The callers cast the
   cumulative block count to int to size .NET collections, so a structural limit
   above int.MaxValue (e.g. from a large AVRO_MAX_COLLECTION_ITEMS override)
   would otherwise reintroduce an int-overflow on that cast.

+++ lang/csharp/src/apache/main/IO/BinaryDecoder.cs
@@ -278,7 +278,9 @@ namespace Avro.IO
             }
 
             EnsureAvailableBytes(p);
-            byte[] buffer = new byte[p];
+            // p has been bounded to <= MaxDotNetArrayLength above, so the cast to
+            // int (required for array allocation) cannot overflow.
+            byte[] buffer = new byte[(int)p];
             Read(buffer, 0, buffer.Length);
             return buffer;
         }
